Combine page entity icon and status icon into one component
