+++ client-v2/src/app/components/atoms/icons/page-entity-icon/page-entity-icon.component.html
@@ -1,3 +1,1 @@
-<status-icon *ngIf="icon != 'tasklist'" [status]="icon"> </status-icon>
-
-<app-icon *ngIf="icon == 'tasklist'" iconClass="far fa-tasks text-tinted-300"></app-icon>
+<app-icon [iconClass]="entityIconClassMap[icon]"></app-icon>
