--- client-v2/src/app/components/atoms/icons/status-icon/status-icon.component.html
@@ -1,1 +1,0 @@
-<app-icon [class]="statusIconClassMap[status]"></app-icon>
