+++ client-v2/src/app/components/organisms/task/task.component.html
@@ -15,7 +15,7 @@
                 [appTooltip]="blocked || data.status"
                 [tooltipOptions]="{ avoidPositions: ['right'] }"
             >
-                <status-icon [status]="loading || blocked || data.status"></status-icon>
+                <page-entity-icon [icon]="loading || blocked || data.status"></page-entity-icon>
             </span>
             <span
                 focusable
